requirements.txt
fastapi==0.101.1
pipreqs==0.4.13
python-dotenv==1.0.0
selenium==4.11.2
uvicorn==0.23.2
webdriver-manager==4.0.0
